+++ src/test/java/com/easypost/ShipmentTest.java
@@ -2,7 +2,6 @@ package com.easypost;
 
 import com.easypost.exception.EasyPostException;
 import com.easypost.exception.General.EndOfPaginationError;
-import com.easypost.exception.General.InvalidParameterError;
 import com.easypost.model.Address;
 import com.easypost.model.EndShipper;
 import com.easypost.model.EstimatedDeliveryDate;
@@ -392,9 +391,6 @@ public final class ShipmentTest {
         assertNotNull(smartRate.getTimeInTransit().getPercentile95());
         assertNotNull(smartRate.getTimeInTransit().getPercentile97());
         assertNotNull(smartRate.getTimeInTransit().getPercentile99());
-
-        assertThrows(InvalidParameterError.class,
-                () -> smartRate.getTimeInTransit().getSmartRateAccuracy("percentile_100"));
     }
 
     /**
@@ -530,20 +526,6 @@ public final class ShipmentTest {
         assertEquals("GroundAdvantage", lowestSmartRateFilters.getService());
         assertEquals(5.93, lowestSmartRateFilters.getRate(), 0.01);
         assertEquals("USPS", lowestSmartRateFilters.getCarrier());
-
-        // Test lowest smartrate with invalid filters (should error due to strict
-        // delivery days)
-        assertThrows(EasyPostException.class, () -> {
-            vcr.client.shipment.lowestSmartRate(shipment.getId(), 0, SmartrateAccuracy.Percentile90);
-        });
-
-        SmartRate deprecatedLowestSmartRateFilters = vcr.client.shipment.lowestSmartRate(shipment.getId(), 3,
-                SmartrateAccuracy.Percentile90);
-
-        // Test lowest smartrate with valid filters
-        assertEquals("GroundAdvantage", deprecatedLowestSmartRateFilters.getService());
-        assertEquals(5.93, deprecatedLowestSmartRateFilters.getRate(), 0.01);
-        assertEquals("USPS", deprecatedLowestSmartRateFilters.getCarrier());
 
         // Test lowest smartrate with invalid filters (should error due to strict
         // delivery days)